Show download status on reader chapter view.
